Make more page
Make a more page. This is starting resemble a tree structure. In fact,
I would really like a tree structure that I get to display. Dirs would
show lists of contents, content files would be rendered. Could have
symlinks and ways to specify collapsing in the rendered view too. Now
add public accessibility as well as file and directory permissions
control and we are talking.